Fix flaky ThreadPoolsExhaustedIntegrationTest
This makes again sure that collect & search share the very same searcher
instance.

We changed that in https://github.com/crate/crate/commit/e7695886044540030b1a50f82f33a52d99c9b1be and since then the ThreadPoolsExhaustedIntegrationTest has been flaky.

Due to the nature of the test, the assumption was that the errors were
related to the kill handling, but logging indicates that the error is
unrelated to the kill handling.